Refactor: Modularize codebase, fix API panics via robust JSON parsing, and add DTLS PSK (-secret) for MitM protection
